Add new team_delete endpoint to handle deleting a team
Also updated team_delete and agent_delete view to return a custom message in the event of delete error related to db integrity. This is to handle a sqlalchemy IntegrityError. Trying to delete an agent or team that has linked players will result in an integrity error due to the nullable=False constraint on the Player model foreign keys. Instead of an http error 500, a custom json message will be returned. Included will be a message outlining the issue, a list of tuples containing player.id and player.name assigned to that agent/team, and the total number of players assigned.